% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_func2arm.R
\name{bayes_pred_go_nogo}
\alias{bayes_pred_go_nogo}
\title{Predicative probabilities and decision tables}
\usage{
bayes_pred_go_nogo(
  nmax = 40,
  nmin = 1,
  p0 = 0.3,
  p1 = 0.5,
  thetats = 0.8,
  thetatf = 0.1,
  thetau = 0.95,
  thetal1 = 0.95,
  a = 1,
  b = 1
)
}
\arguments{
\item{nmax}{max sample size}

\item{nmin}{minimum sample size for the first interim}

\item{p0}{lower reference value of response rate}

\item{p1}{target value of response rate}

\item{thetats}{posterior probability for Final Go s.t. Prob(p > p0) > thetats}

\item{thetatf}{posterior probability for final NoGo Prob(p > p0) <= thetatf}

\item{thetau}{Predictive probability for Go, i.e., PredP(Go) > thetau}

\item{thetal1}{Predictive probability for NoGo PredP(NoGo) > thetal1}

\item{a, b}{priors for beta distribution}
}
\value{
a decision table
}
\description{
Predicative probabilities and decision tables
}
\examples{
go_nogo <- bayes_pred_go_nogo(nmax = 40, nmin = 1, p0 = 0.3, p1 = 0.5,
thetats = 0.8, thetatf = 0.1, thetau = 0.95, thetal1 = 0.95, a = 1, b = 1)

}
