% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-escalation-mtd-estimate.R
\name{boin_decision_matrix}
\alias{boin_decision_matrix}
\alias{boin_tox_interval}
\title{BOIN Decision Matrix}
\usage{
boin_decision_matrix(
  cocap,
  target,
  phi1 = 0.6 * target,
  phi2 = 1.4 * target,
  tox = 0.95
)

boin_tox_interval(target, phi1 = 0.6 * target, phi2 = 1.4 * target)
}
\arguments{
\item{cocap}{cohort cap for each dose}

\item{target}{target toxicity probability}

\item{phi1}{the highest toxicity probability deemed subtherapeutic such that
dose-escalation should be made}

\item{phi2}{lowest toxicity rate deemed overly toxic such that dose
de-escalation is required}

\item{tox}{Unacceptable Toxicity: Prob(Overdosing)}
}
\value{
for \code{boin_decision_matrix}, it returns a decision matrix with
rows being the number of toxicities, and column the corresponding number of
subjects in that cohort

for \code{boin_tox_interval}, it returns a vector of length 2,
corresponding to boundaries for escalation/de-escalation
}
\description{
Derive decision matrix with BOIN method
\loadmathjax
}
\details{
The decision matrix is derived by BOIN method. Given the target
toxicity probability, it first calculates two quantities \mjseqn{\lambda_e}
and \mjseqn{\lambda_d}, in which \code{De-escalation} decision is made if
observed \mjseqn{\hat{p}\in (\lambda_d, 1)} and \code{Stay} decision is recommended when
\mjseqn{\hat{p}\in (\lambda_e, \lambda_d]} and \code{Escalate} when
\mjseqn{\hat{p} \in (0,\lambda_e]} according to
\href{https://www.fda.gov/media/155364/download}{FDA FFP-designated version of BOIN}.
In the original BOIN paper, the boundary values are set as \mjseqn{(0, \lambda_e],
(\lambda_e, \lambda_d)} and \mjseqn{[\lambda_d, 1)}.
}
\examples{
# to repeat the supplementary table S1 in BOIN paper
target1 <- 0.25
x1 <- boin_decision_matrix(18, target1)
# to reproduce the results in S3 of the supplementary table in the BOIN paper
 boin_tox_interval(target = 0.1, phi2 = 1.2*0.1)
}
\references{
{
\insertRef{bayesyuany2016}{r4ct}
}
}
