% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utility.R
\name{bvn_prob}
\alias{bvn_prob}
\title{upper tail probability of a standard binormal distribution}
\usage{
bvn_prob(h, k, rho)
}
\arguments{
\item{h}{the cutoff value for x; will integrate from `h` to `Inf`}

\item{k}{the cutoff value for y; will integrate from `k` to `Inf`}

\item{rho}{the correlation of the bivariate normal distribution.}
}
\value{
a p value
}
\description{
calculate the bivariate normal probabilities for a given h,k and rho
}
\details{
This function calculates the following quantity, \deqn{L(h, k, \rho)
  =\frac{1}{\sqrt{1-\rho^2}}\int_h^{\infty}\int_k^{\infty}e^{-\frac{x^2-2\rho
  xy + y^2}{2(1-\rho^2)}}dxdy} using formula Eq.(3) in the reference.
}
\references{
{
  \insertRef{genz2004numerical}{MethodDev}
}
 \url{https://link.springer.com/content/pdf/10.1023/B:STCO.0000035304.20635.31.pdf}
}
