% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_predictive.R
\name{calc_p1_or}
\alias{calc_p1_or}
\title{odds ratio and probabilities}
\usage{
calc_p1_or(p0, p1 = NULL, or = NULL)
}
\arguments{
\item{p0, p1}{probability in the control/treatment arm}

\item{or}{odds ratio which is expressed as \code{odds_treatment/odds_control}}
}
\value{
probability in the treatment arm, or odds ratio
}
\description{
odds ratio and probabilities
}
\details{
this two functions calculates odds ratio based on two
  probabilities, or probability of treatment given odds ratio and probability
  in the control arm
}
\examples{
 p0 <- 0.59; p1 <- 0.812; or <- 3;
 calc_p1_or(p0 = p0, or = or);
 calc_p1_or(p0 = p0, p1 = p1)
}
