% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_data.R
\docType{data}
\name{click_data}
\alias{click_data}
\title{Click data from the survival curve}
\format{
A data frame with 248 rows and 3 variables:
\itemize{
\item \emph{time}: time along the axis of the survival curve (starting from 0)
\item \emph{survival}: survival probability by the time
\item \emph{arm}: which arm the survival data is from
}
}
\usage{
click_data
}
\description{
A dataset containing the genders and outcomes of two
treatment groups of 100 patients.
}
\keyword{datasets}
