% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{compute_lambda}
\alias{compute_lambda}
\title{Calculate \eqn{\lambda}'s based on the formula}
\usage{
compute_lambda(os, osp, pfs)
}
\arguments{
\item{os}{median os}

\item{osp}{median OS after progression}

\item{pfs}{median PFS}
}
\value{
a vector of lambda's and the theoretical correlation between OS and PFS
}
\description{
Calculate \eqn{\lambda}'s based on the formula
}
