% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cong_chen_2018.R
\name{cong_dat_gen}
\alias{cong_dat_gen}
\title{Simulate data for all comers/enrichment population}
\usage{
cong_dat_gen(nsbj, alloc = c(1, 1), b_size = 2, rate,
  enrichment = FALSE, marker_prob = c(0.7, 0.3),
  marker_name = c("DLL3+", "DLL3-"), par_ctrl_pos, par_ctrl_neg,
  par_trt_pos, par_trt_neg, ...)
}
\arguments{
\item{nsbj}{number of subjects to be simulated}

\item{alloc}{allocation vector, length corresponds to number of arms; as to
be integer; enter 1 if single arm}

\item{b_size}{block size, has to be multiple of sum(alloc), enter 1 if single
arm}

\item{rate}{enrollment rate per unit time}

\item{enrichment}{Is this data generated for enrichment population?}

\item{marker_prob}{vector of prevalence probability of different category, if
doesn't add up to 1, will automatically standardize and generates warning.
For enrichment population, just set \code{marker_prob = 1}}

\item{marker_name}{vector of names of different subgroup}

\item{par_trt_pos, par_trt_neg, par_ctrl_pos, par_ctrl_neg}{parameter
specification for treatment/control and biomarker positive/negative
population. For enrichment data generation, \code{par_trt_neg} and
\code{par_ctrl_neg} are set to be \code{NULL}}

\item{...}{other parameters from function \code{\link[baseUtility]{enrl_dat_gen}}}
}
\value{
a tibble of simulated survival data
}
\description{
Simulate data for all comers/enrichment population
}
