% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utility.R
\name{cong_final_analysis}
\alias{cong_final_analysis}
\title{final analysis for Cong's method}
\usage{
cong_final_analysis(snapshot, marker_positive = "DLL3+",
  alpha1 = 0.0125, alpha2 = 0.0125)
}
\arguments{
\item{snapshot}{the data snapshot}

\item{marker_positive}{the label for biomarker positive population}

\item{alpha1}{significance level for testing all-comers population}

\item{alpha2}{significance level for testing biomarker positive population}
}
\value{
a tibble summarizing the analysis results
}
\description{
final analysis for Cong's method
}
