% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-escalation-mtd-estimate.R
\name{estimate_dlt_isoreg}
\alias{estimate_dlt_isoreg}
\title{Given number of subjects and DLTs, estimate the MTD by isotonic regression
and empirical estimation}
\usage{
estimate_dlt_isoreg(cohort_size, n_dlt, target, du, method = "selfCoded")
}
\arguments{
\item{cohort_size}{a vector of number of subjects in each cohort}

\item{n_dlt}{a vector of number of DLTs (must have the same length as
\code{cohort_size})}

\item{target}{the target toxicity}

\item{du}{a vector with elements 1 or 0 recording whether DU is observed in
the corresponding dose (1) or not (0)}

\item{method}{The method used to perform isotonic regression estimate, either
\code{selfCoded} for  self-implementation or \code{statsPackage} using
\code{\link[stats]{isoreg}}. No matter what method you choose, the results
should be identical.}
}
\value{
a table of results for the estimates
}
\description{
Given number of subjects and DLTs, estimate the MTD by isotonic regression
and empirical estimation
}
\examples{
 n_dlt <- c(0,1,0,0,0,2); cohort_size <- c(3,6,3,6,3,6); du <- c(0, 0, 0, 0, 0, 0)
estimate_dlt_isoreg(cohort_size, n_dlt, du, target = 0.3)
 # it also allows cohort size of 0 in a given cohort
  n_dlt <- c(0,1,0,0,0, 0, 2); cohort_size <- c(3,6,3,0,6,3,6); du <- c(0, 0, 0, 0, 0, 0, 1)
estimate_dlt_isoreg(cohort_size, n_dlt, du, target = 0.3)
}
