% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrington-manning.R
\name{farrington_manning_n}
\alias{farrington_manning_n}
\title{Farrington-Manning sample size calculation}
\usage{
farrington_manning_n(
  p1,
  p2,
  theta,
  delta = NULL,
  r0 = NULL,
  metric = "riskdiff",
  alpha = 0.05,
  beta = 0.2,
  alternative = "greater"
)
}
\arguments{
\item{p1}{response probability in group 1}

\item{p2}{response probability in group 2}

\item{theta}{randomization ratio (in the form of N2/N1)}

\item{delta}{the non-inferiority margin for risk difference under the null}

\item{r0}{the non-inferiority margin for the relative risk under the null}

\item{metric}{specifying the metric used to construct the test statistic,
\code{riskdiff} for testing of risk difference (p1-p2) and \code{relriks} for testing
relative risk (R = p1/p2)}

\item{alpha}{type 1 error control}

\item{beta}{1- power}

\item{alternative}{taking value of either \code{greater} for \code{delta <= delta0} vs
\code{delta > delta0} or \code{two.sided} for \code{delta = delta0} vs \code{delta != delta0}.
Similar for testing relative risk}
}
\value{
a table with sample size for each arm, and p1_tilt, p2_tilt that
estimate p1 and p2 under the null hypothesis.
}
\description{
Farrington-Manning sample size calculation
}
\details{
this function is an implementation of Eq.(4) in Farrington &
Manning, 1990 paper. The sample size calculation is based on \eqn{\hat{p}_1
  - \hat{p}_2 - \delta}, and the test has form
\deqn{H_0: p_1 - p_2 >=\delta VS H_1: p_1-p_2 < \delta } ,
based on maximum likelihood estimation (Method 3 of that paper) under the
null hypothesis restriction \code{p1_tilt - p2_tilt = delta}

\itemize{
\item{p10:}{group 1 proportion tested by the null }
\item{delta:}{non-inferiority margin}
\item{p1:}{binomial proportions = n11/n1}
\item{p2:}{binomial proportions = n21/n2}
\item{p:}{overall proportion  = m1/n}
\item{test} {H0: p10 - p2 >= delta  vs  H1: p10 - p2 < delta}
}

Other forms of hypothesis test are also available. For more details please
refer to the paper.
}
\examples{
# reproducing first row of Table 1 in that paper
farrington_manning_n(p1 = 0.1, p2 = 0.1, r0 = 0.1, theta = 2/3, 
               metric = "relrisk", alpha = 0.05, beta = 0.1)
farrington_manning_n(p1 = 0.1, p2 = 0.1, delta = -0.2, theta = 2/3, 
               metric = "riskdiff", alpha = 0.05, beta = 0.1)
}
\references{
{
\insertRef{farrington1990test}{r4ct}
}
}
