% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrington-manning.R
\name{farrington_manning_significance}
\alias{farrington_manning_significance}
\title{Significance level calculation using Farrington & Manning method}
\usage{
farrington_manning_significance(
  n1,
  n2,
  delta,
  p2,
  alpha = 0.05,
  method = "exact"
)
}
\arguments{
\item{n1}{sample size in group 1}

\item{n2}{sample size in group 2}

\item{delta}{non-inferiority margin}

\item{p2}{the hypothesized response rate in group 2}

\item{alpha}{desired alpha level, e.g. 0.05}

\item{method}{one of \code{exact} or \code{normal}
\itemize{
\item{\code{exact}}{ the test statistics are calculated using exact binomial distribution}
\item{\code{normal}}{ the test statistics are derived based on normal approximation}
}}
}
\value{
numerical value of actual significance level.
}
\description{
Significance level calculation using Farrington & Manning method
}
\examples{
farrington_manning_significance(n1 = 50, n2 = 50, delta = 0.2, p2 = 0.7, alpha = 0.05)
}
