% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrington-manning.R
\name{farrington_manning_z}
\alias{farrington_manning_z}
\title{Farrington & Manning test statistic}
\usage{
farrington_manning_z(
  x1,
  x2,
  n1,
  n2,
  delta = NULL,
  r0 = NULL,
  metric = "riskdiff"
)
}
\arguments{
\item{x1}{the number of events in treatment}

\item{x2}{the nubmer of events in control}

\item{n1}{total number of events in treatment}

\item{n2}{total number of events in control}

\item{delta}{non-inferiority margin for \verb{risk difference}}

\item{r0}{non-inferiority margin for \verb{relative risk}}

\item{metric}{specifying the metric used to construct the test statistic,
\code{riskdiff} for testing of risk difference (p1-p2) and \code{relriks} for testing
relative risk (R = p1/p2)}
}
\value{
the estimated p1_tilt, p2_tilt and test statistic
}
\description{
Farrington & Manning test statistic
}
\details{
test statistics by equation 1 of Farrington and Manning, 1990
this statistic is calculated based on the hypothesis that
H0: p1-p2 >= delta    vs    H1: p1 - p2 < delta
for the cases of (x1, x2) taking value of (0, 0), (0, 1), (1, 0), or (1, 1)
the boundary z values are handled based on Chan's 1999 paper
}
\examples{
 farrington_manning_z(x1 = 10, x2 = 15, n1 = 40, n2 = 40, delta = 0.2)
 farrington_manning_z(x1 = 10, x2 = 15, n1 = 40, n2 = 40, r0 = 0.15, metric = "relrisk")
}
\references{
{
\insertRef{chan1999test}{r4ct}
}
}
