% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-reconstruct-km.R
\name{format_raw_tabs}
\alias{format_raw_tabs}
\title{Augment a raw number at risk table with the necessary information to run
the reconstruction algorithm.}
\usage{
format_raw_tabs(raw_nar, raw_surv, tau = NULL)
}
\arguments{
\item{raw_nar}{A data frame with the columns 'time' and nar' at least.}

\item{raw_surv}{A data frame with the columns 'time' and 'survival' at least.}

\item{tau}{End of follow-up time, defaults to last time in nar table.}
}
\value{
An augmented tab that can be used as input in \code{km_reconstruct}.
}
\description{
Augment a raw number at risk table with the necessary information to run
the reconstruction algorithm.
}
\keyword{internal}
