% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{gen_ospfs}
\alias{gen_ospfs}
\title{generate correlated endpoints of PFS and OS}
\usage{
gen_ospfs(nsbj, mos, mosp = NA, mpfs)
}
\arguments{
\item{nsbj}{number of subjects to be simulated}

\item{mos}{median overall survival time}

\item{mosp}{median overall survival since progression. If \code{mosp = NA},
then OS and PFS will be generated using Theorem 1. If a numerical value is
assigned to \code{mosp}, then OS and PFS will be generated using Theorem 2.}

\item{mpfs}{median PFS}
}
\value{
a tibble with correlated PFS and OS
}
\description{
generate correlated endpoints of PFS and OS
}
\details{
This function generates correlated PFS and OS based on a paper
  published in 2009. Specifically, Let
  \eqn{X_1 \sim \exp(\lambda_1), X_2 \sim \exp(\lambda_2), X_3 \sim \exp(\lambda_3)},
  where \eqn{X_i} has pdf \eqn{f_{X}(x_i) = \lambda_i\exp(-\lambda_ix_i)}.
  \itemize{
   \item{Theorem 1.}{ If \eqn{PFS = \min(X_1, X_2)}, \eqn{OS = X_2} then \deqn{Corr(PFS, OS) = \frac{\lambda_2}{\lambda_1 + \lambda_2}}}
   \item{Theorem 2.}{ If we define \eqn{PFS = \min(X_1, X_2)}, \eqn{OS = PFS} if \eqn{PFS = X_2} and \eqn{OS = X_1 + X_3} otherwise, then \deqn{Corr(PFS, OS) = \frac{\lambda_3}{\sqrt{\lambda_1^2 + 2\lambda_1\lambda_2 + \lambda_3^2}} }}
  }
}
\examples{
mos <- 6.7; mosp  <- 5.4; mpfs <- 1.93
y1 <- gen_ospfs(nsbj = 1e5, mos = mos, mosp = mosp, mpfs = mpfs)
cor(y1)
y2 <- gen_ospfs(nsbj = 1e5, mos, mosp = NA, mpfs)
all.equal(cor(y2)[1,2],mpfs/mos, tolerance = 1e-3)
}
\references{
{
  \insertRef{fleischer2009statistical}{MethodDev}
}
}
