% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-useful-funs.R
\name{get_tte_boundary}
\alias{get_tte_boundary}
\title{For a given event size and randomization ratio, calculate the boundaries for
claiming significance. This is especially useful for group sequential trial
with time-to-event endpoint.}
\usage{
get_tte_boundary(hr_bound, zval, alpha, n_event, rand_ratio)
}
\arguments{
\item{hr_bound}{the HR boundary, significance is declared if observed HR <
\code{hr_bound}}

\item{zval}{the boundary at z-value level}

\item{alpha}{the alpha level (if p value < alpha, significant)}

\item{n_event}{number of events considered for the analysis}

\item{rand_ratio}{randomization ratio in favor of treatment group}
}
\description{
For a given event size and randomization ratio, calculate the boundaries for
claiming significance. This is especially useful for group sequential trial
with time-to-event endpoint.
}
\examples{
get_tte_boundary(hr_bound = NA, zval = NA, alpha = design1$stageLevels[1], 
                 n_event = 199.8, rand_ratio = 1)
# the results should match the following from rpact
# design1 <- getDesignGroupSequential(
# kMax = 2, alpha = 0.025, beta = 0.1, 
# informationRates = c(0.6, 1), typeOfDesign = "asOF")
# getSampleSizeSurvival(design = design1, thetaH0 = 1, 
#                       pi2 = 0.05, hazardRatio = 0.7 )
# NOT RUN
}
