% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-pp-fun.R
\name{gngplot}
\alias{gngplot}
\title{Plot the decision matrix}
\usage{
gngplot(pptabx, groupsize)
}
\arguments{
\item{pptabx}{the result returned by \code{\link{bayes_pred_go_nogo}}}

\item{groupsize}{additional number of subjects needed for each interim after
the first interim}
}
\value{
a plot of decision matrix (ggplot object)
}
\description{
Plot the decision matrix
}
\examples{
go_nogo <- bayes_pred_go_nogo(nmax = 40, nmin = 1, p0 = 0.3, p1 = 0.5,
thetats = 0.8, thetatf = 0.1, thetau = 0.95, thetal1 = 0.95, a = 1, b = 1)
gngplot(go_nogo, groupsize = 5)
}
