% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_histgram}
\alias{hello_histgram}
\title{Create a function for building shiny app packages}
\usage{
hello_histgram(n)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Create a function for building shiny app packages
}
