% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utility.R
\name{ia_pval}
\alias{ia_pval}
\title{Decide interim analysis cutoff value for group sequential design}
\usage{
ia_pval(info_fraction, alpha = 0.025, beta = 0.1, ...)
}
\arguments{
\item{info_fraction}{the information fraction for each look, default set to
be 1, meaning no interim analysis. Can take vector}

\item{alpha}{the desired type 1 error}

\item{beta}{the type 2 error, equal to \code{1 - power}}

\item{...}{other parameters inherited from \code{\link{[gsDesign]gsDesign}}}
}
\value{
a tibble containing number of looks and efficacy/futility p values
}
\description{
Decide interim analysis cutoff value for group sequential design
}
\examples{
ia_pval(alpha = 0.025, beta = 0.1, info_fraction = c(0.5, 0.7))

}
