% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtd-estimate.R
\name{isotonic_reg}
\alias{isotonic_reg}
\title{isotonic regression for estimating probability of MTD}
\usage{
isotonic_reg(x, y)
}
\arguments{
\item{x}{vector of number of toxicities}

\item{y}{vector of number of subject at this dose level}
}
\value{
a vector of toxicity probabilities
}
\description{
isotonic regression for estimating probability of MTD
}
\keyword{internal}
