% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-reconstruct-km.R
\name{km_reconstruct}
\alias{km_reconstruct}
\title{Reconstruct individual-level data from augmented survival table and
nar table, with augmentation performed by \code{format_raw_tabs}.}
\usage{
km_reconstruct(aug_nar, aug_surv)
}
\arguments{
\item{aug_nar}{A data frame processed through \code{format_raw_tabs}}

\item{aug_surv}{A data frame processed through \code{format_raw_tabs}}
}
\value{
A list including \code{ipd_time}, \code{ipd_event}, \code{n_hat = n_hat},
\code{km_hat}, \code{n_cen}, \code{n_event}, \code{int_censor}
}
\description{
Reconstruct individual-level data from augmented survival table and
nar table, with augmentation performed by \code{format_raw_tabs}.
}
\keyword{internal}
