% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_apps.R
\name{launch_app}
\alias{launch_app}
\title{Run the shiny app with an example}
\usage{
launch_app(appname)
}
\arguments{
\item{appname}{the example shiny app to be launched}
}
\value{
a shiny app interface
}
\description{
Run the shiny app with an example
}
\examples{
 # helpful instructions here https://deanattali.com/2015/04/21/r-package-shiny-app/

}
