% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load data from designated format and folder}
\usage{
load_data(data_folder, data_name, data_format, ...)
}
\arguments{
\item{data_folder}{the path to the data}

\item{data_name}{the name of data to be loaded}

\item{data_format}{the format of data, currently support
\itemize{
\item{\code{sas7bdat}} {SAS format with \code{haven::read_sas}}
\item{\code{csv}} {CSV file supported by \code{readr::read_csv}}
\item{\code{xlsx} or \code{xls}} {Excel sheet supported by \code{readxl::read_excel}}
\item{\code{rds}} {rds file by \code{readr::read_rds}}
}}

\item{...}{other parameters}
}
\value{
the data set with column names supported in R
}
\description{
Load data from designated format and folder
}
