% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose-escalation-mtd-estimate.R
\name{mtpi2_decision_matrix}
\alias{mtpi2_decision_matrix}
\alias{plot_decision_matrix}
\title{Bayes Decision Table using modified Toxicity Probability interval (mTPI)}
\usage{
mtpi2_decision_matrix(
  cocap,
  target,
  a,
  b,
  tolerance1,
  tolerance2,
  tox,
  method = "mtpi"
)

plot_decision_matrix(dtab)
}
\arguments{
\item{cocap}{Cohort Cap, max number of subjects in a cohort}

\item{target}{Target Toxicity Rate}

\item{a}{parameter for beta distribution prior}

\item{b}{parameter for beta distribution prior}

\item{tolerance1}{Equivalence Radius -(target tox - tolerance1 = lower acceptance value )}

\item{tolerance2}{Equivalence Radius +(target tox - tolerance2 = upper acceptance value)}

\item{tox}{Unacceptable Toxicity: Prob(Overdosing)}

\item{method}{either \code{mtpi} using mTPI method, or \code{mtpi2} using mTPI2 method}

\item{dtab}{a decision matrix with columns being the number of subjects, and
rows representing corresponding DLTs.}
}
\value{
a decision matrix with rows being the number of toxicities, and
column the corresponding number of subjects in that cohort

a ggplot object
}
\description{
Bayes Decision Table using modified Toxicity Probability interval (mTPI)

Plot the decision matrix for DLT
}
\details{
For each cohort size and number of DLTs observed, calculate the
Bayes posterior probability, based on which a decision table is generated
using mTPI2 (or mTPI) method.
}
\examples{
# create decision table in figure 3
x1 <- mtpi2_decision_matrix(cocap = 10, target = 0.3, a = 1, b = 1, 
   tolerance1 = 0.05, tolerance2 = 0.05, tox = 0.95, method = "mtpi")
x2 <- mtpi2_decision_matrix(cocap = 10, target = 0.3, a = 1, b = 1, 
   tolerance1 = 0.05, tolerance2 = 0.05, tox = 0.95, method = "mtpi2")
# create decision table in figure C1 of the mTPI2 paper
x3 <- mtpi2_decision_matrix(cocap = 10, target = 0.1, a = 1, b = 1, 
   tolerance1 = 0.05, tolerance2 = 0.02, tox = 0.95, method = "mtpi")
x4 <- mtpi2_decision_matrix(cocap = 10, target = 0.1, a = 1, b = 1, 
   tolerance1 = 0.05, tolerance2 = 0.02, tox = 0.95, method = "mtpi2")

x1 <- mtpi2_decision_matrix(cocap = 10, target = 0.3, a = 1, b = 1, 
   tolerance1 = 0.05, tolerance2 = 0.05, tox = 0.95, method = "mtpi")
   plot_decision_matrix(x1)
}
\references{
{
\insertRef{guo2017bayesian}{r4ct}
}
}
