% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpi2_fun.R
\name{mtpi2_decision_plot}
\alias{mtpi2_decision_plot}
\title{Bayes Decision Table for mTPI2}
\usage{
mtpi2_decision_plot(nmax, cocap, tolerance1, tolerance2, a, b, tox, target)
}
\arguments{
\item{nmax}{Maximum Sample Size for the Trial}

\item{cocap}{Cohort Cap, max number of subjects in a cohort}

\item{tolerance1}{Equivalence Radius -(target tox - tolerance1 = lower acceptance value )}

\item{tolerance2}{Equivalence Radius +(target tox - tolerance2 = upper acceptance value)}

\item{a}{parameter for beta distribution prior}

\item{b}{parameter for beta distribution prior}

\item{tox}{Unacceptable Toxicity: Prob(Overdosing)}

\item{target}{Target Toxicity Rate}
}
\value{
a decision plot
}
\description{
Bayes Decision Table for mTPI2
}
\details{
For each cohort size and number of DLTs observed, calculate the
Bayes posterior probability, based on which a decision table is generated
using mTPI2 method.
}
\examples{
mtpi2_decision_plot(nmax = 30, cocap = 12, tolerance1 = 0.05, tolerance2 =
0.05, a = 1, b = 1, tox = 0.95, target = 0.3)
}
\references{
{
\insertRef{guo2017bayesian}{shinyapps4clinicaltrial}
}
}
