% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpi2_fun.R
\name{mTPIeval}
\alias{mTPIeval}
\title{Run simu}
\usage{
mTPIeval(
  B = B0,
  scen = 1,
  cohs = cohs0,
  ncohd = ncohd0,
  ncoh = ncoh,
  mxn = mxn,
  pt = pt,
  pt.a = pt1,
  pt.b = pt2,
  tdose,
  nds = nds,
  ncoh1 = ncoh1,
  nds1 = nds1,
  nds2 = nds2,
  Escf = Escf,
  apr = apr,
  bpr = bpr,
  p.ud = p.ud,
  dslv_start = 5
)
}
\arguments{
\item{B}{}

\item{scen}{}

\item{cohs}{}

\item{ncohd}{}

\item{ncoh}{}

\item{mxn}{}

\item{pt}{}

\item{pt.a}{}

\item{pt.b}{}

\item{tdose}{}

\item{nds}{}

\item{ncoh1}{}

\item{nds1}{}

\item{nds2}{}

\item{Escf}{}

\item{apr}{}

\item{bpr}{}

\item{p.ud}{}

\item{dslv_start}{}
}
\description{
Run simu
}
\keyword{internal}
