% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_predictive.R
\name{pos_two_grid}
\alias{pos_two_grid}
\title{Posterior probabilities for given sample size of n0 and n1}
\usage{
pos_two_grid(n0, n1, p0 = 0.25, p1 = 0.538, delta = (p1 - p0)/2,
  ab0 = NULL, ab1 = NULL)
}
\arguments{
\item{n0, n1}{the sample size for control/treatment group}

\item{p0}{the underlying probability of response rate for the control arm}

\item{p1}{the hypothesized ORR for treatment}

\item{delta}{the difference of the two proportions to be detected}
}
\value{
a tibble which contains the calculated probabilities
}
\description{
for a given pair of \code{n0} and \code{n1}, search all the
  possible sample space, calculates posterior probability, the probailities
  under the null and the alternative. This function is used to evaluate the
  power and type 1 error for a given pair of \code{n0} and \code{n1}
}
\details{
For a given pair of \code{n0} and \code{n1}, after specifying
  appropriate prior parameters, it calculates the posterior probability
  \eqn{P(p_1-p_0 >\delta)}, the probability \eqn{P(X_0 = x_0|n_0, p_0)\cdot
  P(X_1 = x_1|n_1, p_0)} under the null, and \eqn{P(x = x_0|n_0, p_0)\cdot
  P(X = x_1|n_1, p_1)}  under the alternative for each pair of of observed
  \eqn{x_0} and \eqn{x_1}. Note that \eqn{X_0,X_1} are assumed to be
  independent and follow binomial distribution.

  The prior for control group, i.e. \eqn{p_0\sim Beta(a_0,b_0)},  are derived
  based on \eqn{a / (a + b) = p_0} and \eqn{a + b = n_0/2} where \eqn{n_0} is
  the sample size for control arm. The prior for treatment group is obtained
  such that \eqn{a_1 + b_1 = 2} and \eqn{a_1 = 2p_0}.
}
\examples{
library(dplyr)
r1 <- pos_two_grid(n0 = 75, n1 = 75, p0 = 0.59, p1= 0.812)
# power
r1 \%>\% filter(prob_post > 0.68) \%>\% select(prob_alt) \%>\% sum
# type 1 error
r1 \%>\% filter(prob_post > 0.68) \%>\% select(prob_null) \%>\% sum

}
\seealso{
\code{\link[baseUtility]{pos_two}}
}
