% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_predictive.R
\name{pos_two_grid_search}
\alias{pos_two_grid_search}
\title{Power and type 1 error calculation by grid search}
\usage{
pos_two_grid_search(p0, p1, ..., n0 = seq(50, 70, by = 5), n1 = n0,
  cutoff = seq(0.05, 0.2, by = 0.05), eval_success = TRUE,
  ncores = NA, ab0 = NULL)
}
\arguments{
\item{p0}{response rate in the control arm}

\item{p1}{the hypothesized ORR for treatment}

\item{...}{other parameters inherited from \code{\link{pos_two_grid}}}

\item{n0}{sample size for control, can be a vector}

\item{n1}{sample size for treatment, must be of the same length as \code{n0}}

\item{cutoff}{the cutoff value (can be a vector) to claim a decision (either
success or failure)}

\item{eval_success}{Is this for evaluating probability of success? If
\code{TRUE}, then it evaluates \deqn{P(p_1-p_0 > \delta) > U;} otherwise
it evaluates \deqn{P(p_1-p_0>\delta)<L,} where \eqn{L} or \eqn{U} correspond
to \code{cutoff}.}

\item{ncores}{number of cores to be used for fast parallel computing. If not
specified, it will use  number of cores available - 1}

\item{ab0}{a data frame or NULL. If a data frame, it should contain, in each
row, the prior for corresponding sample size \code{n0}. If \code{NULL},
then \code{\link{prior_ab}} will be called internally to calculate the
prior.}
}
\value{
a tibble containing each scenario associated with its power and type
  1 error
}
\description{
for a given pair of \code{n0} and \code{n1}, search all the
  possible sample space, calculates posterior probability, the probailities
  under the null and the alternative. This function is used to evaluate the
  power and type 1 error for a given pair of \code{n0} and \code{n1}
}
\details{
For a given pair of \code{n0} and \code{n1}, after specifying
  appropriate prior parameters, it calculates the posterior probability
  \eqn{P(p_1-p_0 >\delta)}, the probability \eqn{P(X_0 = x_0|n_0, p_0)\cdot
  P(X_1 = x_1|n_1, p_0)} under the null, and \eqn{P(x = x_0|n_0, p_0)\cdot
  P(X = x_1|n_1, p_1)}  under the alternative for each pair of of observed
  \eqn{x_0} and \eqn{x_1}. Note that \eqn{X_0,X_1} are assumed to be
  independent and follow binomial distribution.

  The prior for control group, i.e. \eqn{p_0\sim Beta(a_0,b_0)},  are derived
  based on \eqn{a / (a + b) = p_0} and \eqn{a + b = n_0/2} where \eqn{n_0} is
  the sample size for control arm. The prior for treatment group is obtained
  such that \eqn{a_1 + b_1 = 2} and \eqn{a_1 = 2p_0}.
}
\examples{
temp1 <- pos_two_grid_search(p0 = 0.59, p1 = 0.812,
                n0 = seq(50, 80, by =5),
                cutoff = seq(0.5, 0.8, 0.02),
                ncores = NA,
                eval_success = TRUE)
}
\seealso{
\code{\link{pos_two_grid}}
}
