% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_predictive.R
\name{prior_ab}
\alias{prior_ab}
\title{calculate prior parameters for a given beta distribution}
\usage{
prior_ab(n, p)
}
\arguments{
\item{n}{the size of the prior beta distribution \eqn{a + b = n/2}}

\item{p}{the prior mean \eqn{\frac{a}{a + b} = p}}
}
\value{
the parameters \code{a} and \code{b} for \eqn{Beta(a,b)}
}
\description{
calculate prior parameters for a given beta distribution
}
