% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4ct-package.R
\docType{package}
\name{r4ct-package}
\alias{r4ct}
\alias{r4ct-package}
\title{r4ct: R tools for clinical trials}
\description{
This package documents the R shiny tools I develop over the years through my working experience.
}
\author{
\strong{Maintainer}: bling bling \email{rickyb.zh@gmail.com}

}
\keyword{internal}
