% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{rand_arm_rar}
\alias{rand_arm_rar}
\title{Randomized arm given probability;}
\usage{
rand_arm_rar(
  nsbj,
  blocksize,
  prob,
  arm_name = paste("arm", 1:length(prob), sep = "_")
)
}
\arguments{
\item{nsbj}{number of subjects to be randomized}

\item{blocksize}{the block size?}

\item{prob}{a vector of randomization probabilities}

\item{arm_name}{a vector for the name of the arms}
}
\value{
simulated randomized arms
}
\description{
Randomized arm given probability;
}
\examples{
rand_arm_rar(10, blocksize = 5, prob = c(0.5, 0.5))
rand_arm_rar(10, blocksize = 5, prob = 0.5)
rand_arm_rar(10, blocksize = 5, prob = 2)
}
