% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{rand_inout}
\alias{rand_inout}
\title{Generate subject level enrollment/dropout time by piecewise enroll
rate/attrition probability}
\usage{
rand_inout(nsbj, enrl_timecut, enrl_rate, drop_timecut = NA, drop_prob = NA)
}
\arguments{
\item{nsbj}{number of subjects to generate}

\item{enrl_timecut}{a vector (or NA) specifying starting time for enrollment
\itemize{
\item{a vector}{indicating a piecewise enrollment rate; the vector always
starts with 0}
\item{NA      }{indicating a constant enrollment rate}
}}

\item{enrl_rate}{a vector (or a single value) specifying the enrollment rate
at each piece corresponding enrollment rate.}

\item{drop_timecut}{time cutoff for drop out; if \code{drop_timecut = NA}, t
means no dropout is simulated.}

\item{drop_prob}{dropout probability by \code{drop_timecut}}
}
\value{
a tibble with columns of timein and timeout; Note the dropout time
generated is relative to timein, NOT the calendar time.
}
\description{
Generate subject level enrollment/dropout time by piecewise enroll
rate/attrition probability
}
\details{
This function generates enrollment time and dropout time based on
exponential distribution.
\itemize{
\item{piecewise enrollment time}{under exponential distribution, if it's a
piecewise enrollment, here are the steps: (1) for each piece, sequentially
generate a poission number N to reprensent the number of subjects enrolled
for that piece; (2) generate a vector  of length N from exponential
distribution, representing the enrollment time; (3) combine the pieces
together.
}
\item{dropout time}{Currently only support one-piece dropout!  Algorithm:
(1) given the dropout probability and dropout cut time, calcluate the
dropout hazard ratio by \code{dropout_hr <-
  -log(1-drop_prob)/drop_timecut}; (2) generate dropout time from exponential
distribution with \code{rate = dropout_hr}.
}
}
}
\examples{
inout <- rand_inout(nsbj = 300, enrl_timecut = c(0, 3, 7), enrl_rate = c(2, 5, 8),
drop_timecut = 12, drop_prob = 0.1)

}
