% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{rand_timein}
\alias{rand_timein}
\title{Generate subject level enrollment time by piecewise enrollment rate}
\usage{
rand_timein(nsbj, enrl_timecut, enrl_rate)
}
\arguments{
\item{nsbj}{number of subject enrolled}

\item{enrl_timecut}{a vector (or NA) specifying starting time for enrollment
\itemize{
\item{a vector}{indicating a piecewise enrollment rate; the vector always
starts with 0}
\item{NA      }{indicating a constant enrollment rate}
}}

\item{enrl_rate}{a vector (or a single value) specifying the enrollment rate
at each piece corresponding enrollment rate.}
}
\value{
a tibble where the first column is enrollment time, and the second
column indicates the piece sequence
}
\description{
Generate subject level enrollment time by piecewise enrollment rate
}
\examples{
enrl_rate <- c(7, 14, 30)
enrl_timecut <- c(0, 1, 3)
timein1 <- rand_timein(nsbj = 300, enrl_timecut = enrl_timecut, enrl_rate = enrl_rate)

}
\keyword{internal}
