% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-reconstruct-km.R
\name{reconstruct_survival_data}
\alias{reconstruct_survival_data}
\title{a wrapper function to reconstruct subject level data from KM-curve}
\usage{
reconstruct_survival_data(data_click, data_nar)
}
\arguments{
\item{data_click}{the clicked survival probability tables}

\item{data_nar}{the risk table associated with the survival probability data}
}
\value{
a tibble containing
\describe{
\item{\code{arm}}{the treatment arm}
\item{\code{time}}{survival time}
\item{\code{censor}}{event status, 1 = event, 0 = censor}
}
}
\description{
a wrapper function to reconstruct subject level data from KM-curve
}
\examples{
dat = reconstruct_survival_data(data_click = click_data, data_nar = risk_table)
}
\references{
{
\insertRef{guyot2012enhanced}{r4ct}
}
}
