% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_table.R
\docType{data}
\name{risk_table}
\alias{risk_table}
\title{The risk table from that survival curve}
\format{
A data frame with 10 rows and 4 variables:
\itemize{
\item \emph{time}: time along the axis of the survival curve (starting from 0)
\item \emph{Ipilimumab}: Number of subjects at risk from Ipilimumab arm
\item \emph{\verb{Pembrolizumab, Q2W}}: Number of subejcts at risk from \verb{Pembrolizumab, Q2W} arm
\item \emph{\verb{Pembrolizumab, Q3W}}: number of subjects at risk from \verb{Pembrolizumab, Q3W} arm
}
}
\usage{
risk_table
}
\description{
A dataset containing time, and number of subjects at risk
}
\keyword{datasets}
