% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{run_survival}
\alias{run_survival}
\title{Run survival analysis}
\usage{
run_survival(time, censor, arm, control = NA, conf_level)
}
\arguments{
\item{time}{time to event data}

\item{censor}{indicator for censor or event, 0 = censored, 1 = event}

\item{arm}{indicator for arm. Has to be at least two arms}

\item{control}{the name of the control arm}

\item{conf_level}{The confidence level to calculate for estimated hazard ratio}
}
\value{
a data frame containing results of sample size, p-value, decision,
number of events, hazard ratio with 95\% confidence interval,
median survival, median follow-up, ect
}
\description{
Run survival analysis
}
\examples{
rand_arm(nsbj = 1, ratio = c(1, 1))
time <- c(5.68, 0.34, 4.94, 1.49, 4.72, 1.32, 3.48, 3.42, 3.41, 2.93)
censor <- c(0, 0, 0, 1, 0, 1, 0, 0, 0, 1)
arm <- c("arm_1", "arm_2", "arm_1", "arm_2", "arm_2", "arm_1", "arm_2", "arm_1", "arm_1", "arm_2")
control <- "arm_1"
res1 <- run_survival(time, censor, arm, control = "arm_1", conf_level = 0.95)

}
