% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapps4clinicaltrial-package.R
\docType{package}
\name{shinyapps4clinicaltrial-package}
\alias{shinyapps4clinicaltrial}
\alias{shinyapps4clinicaltrial-package}
\title{shinyapps4clinicaltrial: R shiny tools for clinical trials}
\description{
This package documents the R shiny tools I develop over the years through my working experience.
}
\author{
\strong{Maintainer}: bling bling \email{rickyb.zh@gmail.com}

}
\keyword{internal}
