% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpi2_fun.R
\name{sim_mtpi2}
\alias{sim_mtpi2}
\title{Run the simulation}
\usage{
sim_mtpi2(
  tdose,
  ndose,
  nmax,
  cosize,
  target,
  tolerance1,
  nsim,
  tolerance2,
  cocap,
  tox,
  a,
  b,
  dslv_start
)
}
\arguments{
\item{tdose}{true DLT probabilities}

\item{ndose}{total number of doses}

\item{nmax}{max number of subjects}

\item{cosize}{the cohort size}

\item{target}{target toxicity level}

\item{tolerance1}{Equivalence Radius -(target tox - tolerance1 = lower acceptance value )}

\item{nsim}{number of simulations to run}

\item{tolerance2}{Equivalence Radius +(target tox + tolerance1 = lower acceptance value )}

\item{cocap}{max number of subjects for each cohort}

\item{tox}{Unacceptable Toxicity: Prob(Overdosing)}

\item{a, b}{prior parameters for beta distribution}

\item{dslv_start}{Starting Dose to run simulation}
}
\value{
a list
}
\description{
Run the simulation
}
\examples{
sim_mtpi2(tdose = c(0.05,0.1,0.2,0.25,0.3,0.35), ndose = 6,nmax = 30,
cosize = 3,target = 0.3, tolerance1 = 0.05,nsim = 1000, tolerance2= 0.05, 
cocap = 12,tox = 0.95,  a = 1, b = 1, dslv_start = 3)
}
