% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cong_chen_2018.R
\name{cong_simu_trial}
\alias{cong_simu_trial}
\alias{cong_simu_trial_parallel}
\title{Trial process simulation}
\usage{
cong_simu_trial(n_allcomer, n_enrichment, alloc = c(1, 1), b_size = 2,
  rate, marker_positive = "DLL3+", marker_negative = "DLL3-",
  marker_prob = c(0.7, 0.3), sbj = 100, fu_time_ia = 2,
  n_event = 162, cutoff = 0, alpha1 = 0.0125, alpha2 = 0.0125,
  par_ctrl_pos = list(orr = 0.2, pfs_shape = 1, pfs_median = 7, corr =
  0), par_ctrl_neg = list(orr = 0.1, pfs_shape = 1, pfs_median = 6, corr
  = 0), par_trt_pos = list(orr = 0.1, pfs_shape = 1, pfs_median = 6, corr
  = 0), par_trt_neg = list(orr = 0.3, pfs_shape = 1, pfs_median = 10,
  corr = 0))

cong_simu_trial_parallel(nsim = 10, ncores = 4, ...)
}
\arguments{
\item{n_allcomer}{number of subjects for all comers}

\item{n_enrichment}{increased sample size for enrichment group}

\item{alloc}{allocation vector, length corresponds to number of arms; as to
be integer; enter 1 if single arm}

\item{b_size}{block size, has to be multiple of sum(alloc), enter 1 if single
arm}

\item{rate}{enrollment rate per unit time}

\item{marker_positive, marker_negative}{a string specifying which marker is
negative/positive}

\item{marker_prob}{vector of prevalence probability of different category, if
doesn't add up to 1, will automatically standardize and generates warning}

\item{sbj}{number of subjects for analysis of enrichment decision}

\item{n_event}{the desired number of events for final analysis. Note that
this parameter is used to decide time cutoff for final analysis; therefore
\code{n_event} should be only counted among all-comer populations to proect
the integraty of the trial.}

\item{cutoff}{the cutoff value to determine if enrichment is needed or not}

\item{alpha1, alpha2}{significance level for testing all-comers or biomarker
positive population, respectively}

\item{par_trt_pos, par_trt_neg, par_ctrl_pos, par_ctrl_neg}{parameter
specification for treatment/control and biomarker positive/negative
population}

\item{marker_name}{vector of names of different subgroup}

\item{ia_time_fu}{the follow-up time for decision of enrichment analysis}
}
\description{
This function simulates survival data, performs enrichment
  analysis, based on which results, it enrolls biomarker positive population
  when enrichment is needed, or continues as the original trial when
  enrichment is not needed, and lastly, performs final analysis. Note that
  the final analysis is done on two analysis sets: all-comers and biomarker
  positive population; a win on either population will result in a positive
  outcome.
}
