% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{snapshot_by_event}
\alias{snapshot_by_event}
\title{data snapshot by desired event size}
\usage{
snapshot_by_event(dat, n_event)
}
\arguments{
\item{dat}{the data frame containing, at least, the following variables
\itemize{
\item{\code{timein }}{patient arrival time}
\item{\code{pfs    }}{progression or surivival time}
\item{\code{lfu    }}{lost to follow up time or dropout time}
}}

\item{n_event}{desired number of events for analysis}
}
\value{
the same data with extra columns \code{timecut} (the calander
time cut), \code{pfs_censor} (the censoring indicator, with 1 = event and
0 = censor), \code{ongoing} (whether the status is still ongoing by timecut).
}
\description{
this function calculates the time cut for desired event size and
  then the censor indicator. It has been verified against EAST software.
}
\examples{

}
