% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_test}
\alias{survival_test}
\title{Run survival analysis}
\usage{
survival_test(snapshot, pval_eff = 0.025, pval_fu = NA, is_trt = NA)
}
\arguments{
\item{snapshot}{the data set obtained from
\code{\link[AmgOnc]{take_snapshot}}}

\item{pval_eff, pval_fu}{the significance level to claim a success/futility
for interim analysis or success/failure for final: set \code{pval_eff =
NA} and assign \code{pval_fu} a positive value between 0 and 1 if it's
just for interim futility;  set \code{pval_fu = NA} and assign
\code{pval_eff} a positive value between 0 and 1 if it's just for interim
efficacy; if it's interim analysis for both efficacy and futility, then
must have \code{pval_eff < pval_fu}; if it's for final analysis, then
\code{pval_eff} and \code{pval_fu} must both be specified and set to be
equal;}

\item{is_trt}{user-defined treatment group. If \code{is_trt = NA} then the
second arm number shown in data will be the treatment arm.}
}
\value{
a data frame containing the results of the test
}
\description{
Run survival analysis
}
\details{
this function takes the snapshot and runs survival analysis to get
  the log rank test p-value, the 95% CI for hazard ratio, and the median
  survival time.
}
