% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_predictive.R
\name{tail_prob}
\alias{tail_prob}
\title{Calculate probability of success or failure}
\usage{
tail_prob(dat, cutoff, prob1, prob2, eval_success = TRUE)
}
\arguments{
\item{dat}{the object returned by \code{\link{pos_two_grid}}}

\item{cutoff}{the cutoff value to claim a success/failure}

\item{prob1}{the posterior probability}

\item{prob2}{the probability under the null or the alternative}

\item{eval_success}{Is this for evaluating probability of success?}
}
\value{
a p value
}
\description{
Calculate probability of success or failure
}
