% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utility.R
\name{test_bm_neg}
\alias{test_bm_neg}
\title{Enrichment decision making}
\usage{
test_bm_neg(cong_dat, marker_negative = "DLL3-", endpoint = "resp",
  sbj = 100, fu_time_ia = 2, cutoff = 0)
}
\arguments{
\item{cong_dat}{a data set generated by \code{\link{cong_dat_gen}}}

\item{marker_negative}{a string specifying which marker is negative}

\item{endpoint}{the endpoint used to calculate the decision rule}

\item{sbj}{number of subjects to be included in analysis of biomarker negative}

\item{fu_time_ia}{minimum follow-up time for eligible evaluation}

\item{cutoff}{the cutoff chosen to make the decision}
}
\value{
a tibble with decision included (see column \code{need_enrichment})
}
\description{
Enrichment decision making
}
\details{
This function performs analysis for the biomarker negative
  population, then decides if enrichment is needed
}
