% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtpi2_fun.R
\name{estimate_dlt_isoreg}
\alias{estimate_dlt_isoreg}
\title{Given number of subjects and DLTs, estimate the MTD by isotonic regression
and empirical estimation}
\usage{
estimate_dlt_isoreg(cohort_size, n_dlt, target)
}
\arguments{
\item{cohort_size}{a vector of number of subjects in each cohort}

\item{n_dlt}{a vector of number of DLTs (must have the same length as
\code{cohort_size})}

\item{target}{the target toxicity}
}
\value{
a table of results for the estimates
}
\description{
Given number of subjects and DLTs, estimate the MTD by isotonic regression
and empirical estimation
}
\examples{
 n_dlt <- c(0,1,0,0,0,2); cohort_size <- c(3,6,3,6,3,6)
estimate_dlt_isoreg(cohort_size, n_dlt, target = 0.3)
}
