% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{rand_arm}
\alias{rand_arm}
\title{generate blocked randomized arms}
\usage{
rand_arm(nsbj, ratio)
}
\arguments{
\item{nsbj}{a vector specifying number of subjects to be randominzed for each
arm}

\item{ratio}{the allocation ratio}
}
\value{
a vector of length `sum(nsbj)` with randomized treatment arms
}
\description{
generate blocked randomized arms
}
\examples{
rand_arm(nsbj = c(0, 1), ratio = c(1, 1))
rand_arm(nsbj = c(5, 5, 2), ratio = c(2, 2, 1))

}
