% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{rand_arm}
\alias{rand_arm}
\title{generate block randomized arms}
\usage{
rand_arm(nsbj, ratio, arm_name = paste("arm", 1:length(ratio), sep =
  "_"))
}
\arguments{
\item{nsbj}{an integer for total number of subjects to be randominzed}

\item{ratio}{the allocation ratio}

\item{arm_name}{a vector of characters for arms}
}
\value{
a vector of length `nsbj` with randomized treatment arms
}
\description{
generate block randomized arms
}
\examples{
rand_arm(nsbj = 1, ratio = c(1, 1))
rand_arm(nsbj = 12, ratio = c(2, 2, 1))
rand_arm(nsbj = 4, ratio = c(1, 2, 0, 1))

}
