% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{rand_timein}
\alias{rand_timein}
\title{Generate enrollment time by piecewise enrollment rate}
\usage{
rand_timein(nsbj, rate, starttime)
}
\arguments{
\item{nsbj}{number of subject enrolled}

\item{rate}{a vector (or a single value) specifying the enrollment rate at
each piece}

\item{starttime}{a vector (or a single value) specifying starting time for
corresponding enrollment rate. \code{starttime} always starts with 0,
whether it's a vector or a single value.}
}
\value{
a tibble where the first column is enrollment time, and the second
  column indicates the piece sequence
}
\description{
Generate enrollment time by piecewise enrollment rate
}
\examples{
rate <- c(7, 14, 30)
starttime <- c(0, 1, 3)
timein1 <- rand_timein(nsbj = 300, rate = rate, starttime = starttime)

}
